+++ src/database/operations/load/index.ts
@@ -3,5 +3,7 @@ export * from './loadBulkUpload';
 export * from './loadBulkUploadBundle';
 export * from './loadBundle';
 export * from './loadObjects';
+export * from './loadOrganization';
+export * from './loadOrganizationBundle';
 export * from './loadProposalBundle';
 export * from './loadTableMetrics';