+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.1 - 2024-02-29
+
+### Added
+
+- The `Organization` entity now exists.
+- The GET `/organizations` endpoint now exists.
+- The GET `/organizations:id` endpoint now exists.
+- The POST `/organizations` endpoint now exists.
+
 ## 0.3.0 - 2023-05-03
 
 ### Changed